Add common block change rewriter
