+++ common/src/main/java/com/viaversion/viabackwards/protocol/protocol1_10to1_11/packets/BlockItemPackets1_11.java
@@ -54,6 +54,9 @@ public class BlockItemPackets1_11 extends LegacyBlockItemRewriter<ClientboundPac
 
     @Override
     protected void registerPackets() {
+        registerBlockChange(ClientboundPackets1_9_3.BLOCK_CHANGE);
+        registerMultiBlockChange(ClientboundPackets1_9_3.MULTI_BLOCK_CHANGE);
+
         protocol.registerClientbound(ClientboundPackets1_9_3.SET_SLOT, new PacketHandlers() {
             @Override
             public void register() {
@@ -184,34 +187,6 @@ public class BlockItemPackets1_11 extends LegacyBlockItemRewriter<ClientboundPac
                 if (id.equals("minecraft:sign")) {
                     idTag.setValue("Sign");
                 }
-            }
-        });
-
-        protocol.registerClientbound(ClientboundPackets1_9_3.BLOCK_CHANGE, new PacketHandlers() {
-            @Override
-            public void register() {
-                map(Type.POSITION1_8); // 0 - Block Position
-                map(Type.VAR_INT); // 1 - Block
-
-                handler(wrapper -> {
-                    int idx = wrapper.get(Type.VAR_INT, 0);
-                    wrapper.set(Type.VAR_INT, 0, handleBlockID(idx));
-                });
-            }
-        });
-
-        protocol.registerClientbound(ClientboundPackets1_9_3.MULTI_BLOCK_CHANGE, new PacketHandlers() {
-            @Override
-            public void register() {
-                map(Type.INT); // 0 - Chunk X
-                map(Type.INT); // 1 - Chunk Z
-                map(Type.BLOCK_CHANGE_RECORD_ARRAY);
-
-                handler(wrapper -> {
-                    for (BlockChangeRecord record : wrapper.get(Type.BLOCK_CHANGE_RECORD_ARRAY, 0)) {
-                        record.setBlockId(handleBlockID(record.getBlockId()));
-                    }
-                });
             }
         });
